+++ src/extension.ts
@@ -8,11 +8,13 @@ import { registerOpenTutorial } from './commands/openTutorial';
 import { registerSearchReactorLibrary } from './community/browser';
 import { registerGeometryPreview } from './preview/webview';
 import { registerSnippetCompletions } from './completions/snippets';
+import { registerHighlightPalettes } from './highlight';
 
 export function activate(context: vscode.ExtensionContext) {
     console.log('OWEN extension activated');
 
     registerSnippetCompletions(context);
+    registerHighlightPalettes(context);
 
     context.subscriptions.push(
         vscode.commands.registerCommand('owen.openLatticeBuilder', () => {